requirements.txt
Django==2.2.17

pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-factoryboy==2.0.3
pytest-mock==3.3.1
pytest-pythonpath==0.7.3